requirements.txt
boto3
fastapi 
uvicorn 
requests
google-genai
httpx
python-dotenv